refactor : 백엔드 유틸 함수 분리
